♻️: refactor: update package name no user uppercase
