+++ src/main/java/com/grimeet/grimeet/domain/auth/service/AuthServiceImpl.java
@@ -14,7 +14,7 @@ import com.grimeet.grimeet.domain.user.dto.UserResponseDto;
 import com.grimeet.grimeet.domain.user.entity.User;
 import com.grimeet.grimeet.domain.user.repository.UserRepository;
 import com.grimeet.grimeet.domain.user.service.UserDetailServiceImpl;
-import com.grimeet.grimeet.domain.userLog.service.UserLogFacade;
+import com.grimeet.grimeet.domain.user_log.service.UserLogFacade;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
