+++ src/main/java/com/grimeet/grimeet/domain/auth/service/GoogleAuthServiceImpl.java
@@ -3,10 +3,9 @@ package com.grimeet.grimeet.domain.auth.service;
 import com.grimeet.grimeet.common.config.oauth.info.GoogleOAuthUserInfo;
 import com.grimeet.grimeet.common.exception.ExceptionStatus;
 import com.grimeet.grimeet.common.exception.GrimeetException;
-import com.grimeet.grimeet.domain.socialAccount.dto.Provider;
-import com.grimeet.grimeet.domain.socialAccount.dto.SocialAccountRequestDto;
-import com.grimeet.grimeet.domain.socialAccount.service.SocialAccountFacade;
-import com.grimeet.grimeet.domain.user.service.UserFacade;
+import com.grimeet.grimeet.domain.social_account.dto.Provider;
+import com.grimeet.grimeet.domain.social_account.dto.SocialAccountRequestDto;
+import com.grimeet.grimeet.domain.social_account.service.SocialAccountFacade;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
