+++ src/main/java/com/grimeet/grimeet/domain/websocket/draw_data/service/DrawDataService.java
@@ -1,6 +1,6 @@
-package com.grimeet.grimeet.domain.websocket.drawData.service;
+package com.grimeet.grimeet.domain.websocket.draw_data.service;
 
-import com.grimeet.grimeet.domain.websocket.drawData.dto.DrawDataRequestDto;
+import com.grimeet.grimeet.domain.websocket.draw_data.dto.DrawDataRequestDto;
 
 public interface DrawDataService {
   void saveDrawData(DrawDataRequestDto requestDto, String userEmail);
